bugfix/ initial load fail
Fixed bug where app won't load and throws errors. Problem was IP-info API. It had this 5/5min (?) limit of requests. I wouldn't be thinking that could be a problem. Now Application is using `navigator` object to get the coordinates and this is used to fetch weather. Of course user needs to agree to share his location, If he won't - Tokyo is used as default location.